UPDATE : 구글 로그인 api 연동 및 prettier 삭제
