--- tmeet-frontend/.eslintrc.json
@@ -1,8 +1,0 @@
-{
-  "extends": ["react-app", "prettier"],
-  "plugins": ["prettier"],
-  "rules": {
-    "prettier/prettier": ["error", { "endOfLine": "auto" }],
-    "react/jsx-filename-extension": 0
-  }
-}
