--- tmeet-frontend/.prettierrc.yaml
@@ -1,7 +1,0 @@
----
-trailingComma: 'es5'
-tabWidth: 2
-semi: true
-singleQuote: true
-printWidth: 120
-endOfLine: lf
