+++ tmeet-frontend/package.json
@@ -7,6 +7,7 @@
     "@emotion/styled": "^11.8.1",
     "@mui/icons-material": "^5.6.2",
     "@mui/material": "^5.6.4",
+    "@react-oauth/google": "^0.2.6",
     "@shannenwu/react-schedule-selector": "^1.0.4",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
@@ -41,10 +42,6 @@
     ]
   },
   "devDependencies": {
-    "eslint": "8.0.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-config-react-app": "7.0.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "prettier": "^2.6.2"
+    "eslint": "8.0.0"
   }
 }
