+++ tmeet-frontend/src/App.js
@@ -6,7 +6,8 @@ import LoginSample from './pages/sample/LoginSample';
 import TimeSample from './pages/sample/TimeSample';
 import Time from './pages/time/Time';
 import TimeSelect from './pages/timeselect/TimeSelect';
-import Test from './pages/login/Test'
+import KakaoRedirectTest from './pages/login/KakaoLogin/KakaoRedirectTest';
+import GoogleLogin from "./pages/login/GoogleLogin/GoogleLogin";
 
 function App() {
   return (
@@ -19,7 +20,8 @@ function App() {
           <Route exact path="/sample/time" element={<TimeSample />}></Route>
           <Route exact path="/time" element={<Time />}></Route>
           <Route exact path="/timeselect" element={<TimeSelect />}></Route>
-          <Route exact path="/test" element={<Test />}></Route>
+          <Route exact path="/kakao-redirect" element={<KakaoRedirectTest />}></Route>
+          <Route exact path="/google-login" element={<GoogleLogin />}></Route>
         </Route>
       </Routes>
     </BrowserRouter>
