--- tmeet-frontend/src/pages/login/Test.jsx
@@ -1,13 +1,0 @@
-import NavBar from "../components/navigationBar/NavBar";
-
-
-function App() {
-    return (
-        <div className="App">
-            <NavBar></NavBar>
-            <h1>test 성공</h1>
-        </div>
-    );
-}
-
-export default App;
